[SL] Reorganize files
