+++ backend/matching-service/server.js
@@ -68,7 +68,7 @@ wss.on("connection", (ws) => {
         else {
           ws.matchedUser = matchedUser;
           matchedUser.matchedUser = ws;
-          const uniqueRoomId = generateCodeRoomId()
+          const uniqueRoomId = generateCodeRoomId();
           ws.send(JSON.stringify({ status: "matched", user_id: matchedUser.userId, room_id: uniqueRoomId }));
           matchedUser.send(JSON.stringify({ status: "matched", user_id: ws.userId, room_id: uniqueRoomId }));
           console.info(`"${complexity}": Matched user ${matchedUser.userId} to user ${userId}`);
