+++ frontend/src/components/CodeRoom/Sidebar/Sidebar.module.css
@@ -1,6 +1,6 @@
 .sidebar_container {
   color: #fff;
-  background-color: #25272a;
+  background-color: #282828;
 }
 
 .sidebar {
